Edit some numerical types

+++ lib.partial_system/src/main/kotlin/core/base/BaseCollision.kt
@@ -3,5 +3,5 @@ package core.base
 abstract class BaseCollision<P : BaseParticle<T, O>, T : BasePoint<O>, O : BaseOffset<T>>(
     open val particle1: P,
     open val particle2: P,
-    open val dist: Float
+    open val dist: Double
 )
