+++ src/main/kotlin/radar/scene/CatCollision.kt
@@ -5,6 +5,6 @@ import core.base.BaseCollision
 data class CatCollision(
     override val particle1: CatParticle,
     override val particle2: CatParticle,
-    override val dist: Float,
+    override val dist: Double,
     val catState: CatStates
 ) : BaseCollision<CatParticle, Point2D, Offset2D>(particle1, particle2, dist)
